+++ Quadriga/Assets/Scripts/Car_Controller.cs
@@ -8,8 +8,11 @@ public class Car_Controller : MonoBehaviour
     public float acceleration;
     public float torquePower;
     public float maxSpeed;
+    public float minSpeed;
     public float breakPower;
     public bool breaking;
+    public bool slowedDown;
+    public bool speededUp;
     Rigidbody2D body;
     void Start()
     {
@@ -18,18 +21,31 @@ public class Car_Controller : MonoBehaviour
     }
 
     private void Update()
-    {
-        Debug.Log(breaking);
-        if (speed < maxSpeed )
-            speed += acceleration;
-        if (Input.GetKey(KeyCode.A) && Input.GetKey(KeyCode.D) && speed > 0.02f)
-        {
+    {     
+        if (Input.GetKey(KeyCode.A) && Input.GetKey(KeyCode.D))
+        {        
             speed -= breakPower;
             breaking = true;
+            if (speed < minSpeed)
+                speed = minSpeed;
         }
         else
+        {
+            if (speed < maxSpeed)
+                speed += acceleration;
+            else speed = maxSpeed;
             breaking = false;
-            
+        }
+        if (speed == minSpeed)
+            slowedDown = true;
+        else
+            slowedDown = false;
+        if (speed == maxSpeed)
+            speededUp = true;
+        else
+            speededUp = false;
+
+
     }
     void FixedUpdate()
     {
